fix delays in signal handling
There is no need to loop in the master_sleep method at all, as
the rest of the code is designed to function even on interrupted
sleeps.

This change is included as part of a larger cleanup in master.
(commit bdc79712e5ac53d39c51e80dfe50aff950e5053f)