http: support keepalive_requests directive
This limits the number of keepalive requests of a single
connection to prevent a single client from monopolizing server
resources.  On multi-process servers (e.g. Rainbows!) with many
keepalive clients per worker process, this can force a client to
reconnect and increase its chances of being accepted on a
less-busy worker process.

This directive is named after the nginx directive which
is identical in function.