shutdown client socket for apps which fork in background
Previously we relied on implicit socket shutdown() from the
close() syscall.  However, some Rack applications fork()
(without calling exec()), creating a potentially long-lived
reference to the underlying socket in a child process.  This
ends up causing nginx to wait on the socket shutdown when the
child process exits.

Calling shutdown() explicitly signals nginx (or whatever client)
that the unicorn worker is done with the socket, regardless of
the number of FD references to the underlying socket in
existence.

This was not an issue for applications which exec() since
FD_CLOEXEC is always set on the client socket.

Thanks to Patrick Wenger for discovering this.  Thanks to
Hongli Lai for the tip on using shutdown() as is done in
Passenger.

ref: http://mid.gmane.org/CAOG6bOTseAPbjU5LYchODqjdF3-Ez4+M8jo-D_D2Wq0jkdc4Rw@mail.gmail.com