README: clarify license and copyright
Since Ruby 1.9.3, (Matz) Ruby is licensed under a 2-clause BSDL.
Thus we need to clarify we inherited the license terms from
Ruby 1.8 to prevent misunderstanding.

(The Ruby license change cannot alter the license of other
 projects automatically)

Since we added the GPLv3 as an additional license in 2011,
the license terms of unicorn no longer matches Mongrel 1.1.5.

This is NOT a change to the unicorn license at all, just a
wording clarification.