Add rack.after_reply functionality
This adds `rack.after_reply` functionality which allows rack middleware
to pass lambdas that will be executed after the client connection has
been closed.

This was driven by a need to perform actions in a request that shouldn't
block the request from completing but also don't make sense as
background jobs.

There is prior art of this being supported found in a few gems, as well
as this functionality existing in other rack based servers (e.g. Puma).

[ew: check if `env' is set in ensure statement]

Acked-by: Eric Wong <e@80x24.org>