Spawn workers from the mold if one exist
Rather than direclty fork from the master, we send
a message to the mold to ask for another worker to be forked.

The new worker will use a double fork to detach itself from
its parent, and will be re-attached to the master becase we
enabled PR_SET_CHILD_SUBREAPER.

If the system doesn't support PR_SET_CHILD_SUBREAPER, we simply
never create a mold.